Add scaled channel configuration dialog
