+++ OpenEphys.Onix1.Design/NeuropixelsV1ChannelConfigurationDialog.cs
@@ -10,7 +10,7 @@ namespace OpenEphys.Onix1.Design
     /// <summary>
     /// Partial class to create a channel configuration GUI for the <see cref="ConfigureNeuropixelsV2e"/> class.
     /// </summary>
-    public partial class NeuropixelsV1ChannelConfigurationDialog : ChannelConfigurationDialog
+    public partial class NeuropixelsV1ChannelConfigurationDialog : ScaledChannelConfigurationDialog
     {
         internal event EventHandler OnZoom;
         internal event EventHandler OnFileLoad;
@@ -90,8 +90,6 @@ namespace OpenEphys.Onix1.Design
         {
             OnZoom?.Invoke(this, EventArgs.Empty);
         }
-
-        internal override bool IsDrawScale() => true;
 
         internal override void DrawScale()
         {
